Moved formatter unit test out of integration tests.
